// frontend/src/client/types.gen.ts
export type ItemPublic = {
    title: string;
    description?: (string | null);
    location?: (string | null);
    price?: (string | null);
    difficulty?: (string | null);
    type?: (string | null);
    category?: (string | null);
    tags?: Array<(string)>;
    link?: (string | null);
    picture?: (string | null);
    in_featured?: boolean;
    rating?: (number | null);
    tries?: (number | null);
    favorites?: (number | null);
    duration?: (string | null);
    id: string;
    owner_id: string;
};